+++ package.json
@@ -560,6 +560,8 @@
         "triple-beam": "^1.3.0",
         "typescript": "^3.7.2",
         "uuid": "^3.3.2",
+        "vscode-debugadapter": "1.38.0",
+        "vscode-debugprotocol": "1.39.0",
         "vscode-nls": "^4.1.1",
         "vue": "^2.5.16",
         "winston": "^3.2.1",
